+++ pkg/subprocess/subprocess_darwin.go
@@ -12,22 +12,22 @@ import (
 type SubprocessStatus int
 
 const (
-	SubprocessStatusNotRunning SubprocessStatus = iota
+	SubprocessStatusNotStarted SubprocessStatus = iota
 	SubprocessStatusRunning
 	SubprocessStatusFinished
 )
 
 type Subprocess struct {
-	Status         SubprocessStatus
-	WDir           string
-	Program        string
-	Args           []string
-	Process        *os.Process
-	StdinStream    io.WriteCloser
-	StdoutCallback func(string)
-	StderrCallback func(string)
-	PreProcess     func()
-	PostProcess    func()
+	Status         SubprocessStatus `json:"status"`
+	WDir           string           `json:"-"`
+	Program        string           `json:"-"`
+	Args           []string         `json:"-"`
+	Process        *os.Process      `json:"-"`
+	StdinStream    io.WriteCloser   `json:"-"`
+	StdoutCallback func(string)     `json:"-"`
+	StderrCallback func(string)     `json:"-"`
+	PreProcess     func()           `json:"-"`
+	PostProcess    func()           `json:"-"`
 }
 
 func (p *Subprocess) Execute() {
@@ -41,6 +41,7 @@ func (p *Subprocess) Execute() {
 	err := c.Start()
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "can't execute subprocess: %s\n", err)
+		return
 	}
 
 	p.Status = SubprocessStatusRunning